refactoring: putting interfaces in own file
